Changed class and namespace names.
