+++ Chummer/Chummer.csproj
@@ -280,9 +280,10 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Backend\Attributes\AttributeSection.cs" />
-    <Compile Include="Backend\BuySellIncreaseDecreaseMethods\ChangeAttributes.cs" />
-    <Compile Include="Backend\BuySellIncreaseDecreaseMethods\ChangeCyberware.cs" />
-    <Compile Include="Backend\BuySellIncreaseDecreaseMethods\ChangeMetamagicOrEcho.cs" />
+    <Compile Include="Backend\StaticMethods\StaticAttributes.cs" />
+    <Compile Include="Backend\StaticMethods\StaticCyberware.cs" />
+    <Compile Include="Backend\StaticMethods\StaticMetamagicAndEcho.cs" />
+    <Compile Include="Backend\StaticMethods\CharacterCalculations.cs" />
     <Compile Include="Backend\Datastructures\AvailabilityValue.cs" />
     <Compile Include="Backend\Datastructures\CachedBindingList.cs" />
     <Compile Include="Backend\Datastructures\ConcurrentHashSet.cs" />
